Added DeleteFilter to Syncer struct

+++ fsync.go
@@ -55,6 +55,9 @@ type Syncer struct {
 	// Set this to true to delete files in the destination that don't exist
 	// in the source.
 	Delete bool
+	// To allow certain files to remain in the destination, implement this function.
+	// Return true to skip file, false to delete.
+	DeleteFilter func(f os.FileInfo) bool
 	// By default, modification times are synced. This can be turned off by
 	// setting this to true.
 	NoTimes bool
@@ -63,7 +66,11 @@ type Syncer struct {
 
 // NewSyncer creates a new instance of Syncer with default options.
 func NewSyncer() *Syncer {
-	return &Syncer{}
+	s := Syncer{}
+	s.DeleteFilter = func(f os.FileInfo) bool {
+		return false
+	}
+	return &s
 }
 
 // Sync copies files and directories inside src into dst.
@@ -182,6 +189,9 @@ func (s *Syncer) sync(dst, src string) {
 		files, err = ioutil.ReadDir(dst)
 		check(err)
 		for _, file := range files {
+			if s.DeleteFilter(file) {
+				continue
+			}
 			if !m[file.Name()] {
 				check(os.RemoveAll(filepath.Join(dst, file.Name())))
 			}
